Add build strategy
In the build process, the creation of stable versions is pursued. So-called release tags are used for this. Release tags are not a feature supported by Git; instead, the commit ID of the main branch linked to a release tag is used.

The intention or necessity of a multi-stage build process is to map existing and new aspects as a coherent chain.

The first step is to make modifications to the base module of the original project so that it recognises a new parent module (template).

If the project to be integrated has a parent module, this is first replaced by a new parent module (template). Afterwards the new template is instructed to recognise the existing parent module of the original project.

The second step is to specify the version number of the project to be used. The detected version number is set in the new parent module (template).

Components / files: settings-modify.xml, pom-modify.xml, pom.xml

The final step is the construction of the overall project.

Components / files: settings.xml, pom.xml